requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
antlr4-python3-runtime==4.8
    # via
    #   hydra-core
    #   omegaconf
cycler==0.10.0
    # via matplotlib
hydra-core==1.1.1
    # via
    #   -r requirements.in
    #   hydra-experiment-sweeper
hydra-experiment-sweeper==0.1.0
    # via -r requirements.in
kiwisolver==1.3.2
    # via matplotlib
matplotlib==3.4.3
    # via -r requirements.in
more-itertools==8.10.0
    # via hydra-experiment-sweeper
numpy==1.21.3
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
    #   scipy
omegaconf==2.1.1
    # via hydra-core
pandas==1.3.4
    # via -r requirements.in
pillow==8.4.0
    # via matplotlib
pyparsing==3.0.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2021.3
    # via pandas
pyyaml==6.0
    # via omegaconf
scipy==1.7.1
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
